Task 4.21 Blog List Expansion, step9

+++ part4/bloglistbackend/controllers/blogs.js
@@ -53,6 +53,24 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!decodedToken.id) {
+    return response.status(401).json({error: 'token invalid'})
+  }
+
+  const loggedInUser = await User.findById(decodedToken.id)
+  if(!loggedInUser) {
+    return response.status(400).json({error: 'UserId missing or not valid'})
+  }
+
+  const blog = await Blog.findById(request.params.id)
+  if (!blog) {
+    return response.status(404).json({ error: 'blog not found' })
+  }
+
+  if(blog.user.toString() !== loggedInUser.id.toString()) {
+    return response.status(403).json({ error: 'unauthorized: not the blog creator' })
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
